feat(tests): Stand up tests
Additions:
* (pyproject.toml): Add a `whisperlab` entry point
* (tests/data/hello_world.mp3): Add a new test sample
* (logging): Add a `debug` option
* (pyproject.toml): Use a single line verbose format

Refactors:
* (__init__): Move version logic into __init__

Fixes:
* (__main__): Tasks/Pydantic models require keyword arguments
* (test_main): Use a custom runner (See: Note 1)
* (test_main::test_transcribe): Test the "hello_world.mp3" sample
* (transcribe): Eliminate fp16 warning
* (test_pypath): Diagnose VSCode test runner issue (See: Note 2)


# Note 1: 

Click's CliRunner.invoke throws a file not found error when running test cases that invoke the log. This appears to be a race or other contention between CliRunner and pytest's live logging.

To work around this, I've built a run_whisperlab function based on subprocess.


# Note 2:

VSCode's test runner fails to add venv/bin to the path. This causes `subprocess.run` to throw a `FileNotFound` error when running the `whisperlab` binary

To work around this issue we can:

1. Call pytest from the CLI
2. Ignore the test_main module
3. Monkeypatch the path in the test base module
4. Add custom vscode settings or env
5. Preferred: Diagnose and fix the test runner issue